pyproject.toml
[project]
name = "notion-reddit-analysis"
version = "0.1.0"
description = "Automated competitive intelligence tool for analyzing r/Notion subreddit"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "fake-useragent>=1.4.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "wordcloud>=1.9.0",
    "textblob>=0.17.0",
    "aiohttp>=3.8.0",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "google-generativeai>=0.3.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
api = [
    "praw>=7.7.1",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"